+++ projects/feather-theme/src/public-api.ts
@@ -6,3 +6,5 @@ export * from './lib/theme.module';
 export * from './lib/theme.service';
 export * from './lib/models/theme.model';
 export * from './lib/directives/theme.directive';
+export { FtMessageService } from './services/ft-message.service';
+export type { MessageType, ToastButton, ToastOptions } from './services/ft-message.service';